pin_project_internal

Module utils

Macros

parse_quote_spanned 🔒

Structs

ReplaceReceiver 🔒

Traits

ParseBufferExt 🔒

SliceExt 🔒

Functions

contains_fn 🔒

determine_lifetime_name 🔒

Determines the lifetime names. Ensure it doesn’t overlap with any existing lifetime names.

determine_visibility 🔒

Determines the visibility of the projected types and projection methods.

insert_lifetime 🔒

Inserts a lifetime at position 0 of generics.params.

insert_lifetime_and_bound 🔒

Like insert_lifetime, but also generates a bound of the form OriginalType<A, B>: 'lifetime. Used when generating the definition of a projection type

prepend_underscore_to_self 🔒

respan 🔒

respan_tokens 🔒

Type Aliases

Variants 🔒