pin_project_internal::utils

Function determine_visibility

pub(crate) fn determine_visibility(vis: &Visibility) -> Visibility

Determines the visibility of the projected types and projection methods.

If given visibility is pub, returned visibility is pub(crate). Otherwise, returned visibility is the same as given visibility.