pin_project_internal::utils

Function insert_lifetime_and_bound

pub(crate) fn insert_lifetime_and_bound(
    generics: &mut Generics,
    lifetime: Lifetime,
    orig_generics: &Generics,
    orig_ident: &Ident,
) -> WherePredicate

Like insert_lifetime, but also generates a bound of the form OriginalType<A, B>: 'lifetime. Used when generating the definition of a projection type