Module pin_project_internal::utils

Structs

ReplaceReceiver

Enums

ProjKind

Traits

ParseBufferExt

SliceExt

VecExt

Functions

contains_fn

determine_lifetime_name
Determines the lifetime names. Ensure it doesn't overlap with any existing lifetime names.

determine_visibility
Determines the visibility of the projected type and projection method.

insert_lifetime
Inserts a lifetime at position 0 of generics.params.

insert_lifetime_and_bound
Like insert_lifetime, but also generates a bound of the form OriginalType<A, B>: 'lifetime. Used when generating the definition of a projection type

parse_as_empty
Check if tokens is an empty TokenStream. This is almost equivalent to syn::parse2::<Nothing>(), but produces a better error message and does not require ownership of tokens.

prepend_underscore_to_self

respan

respan_tokens

Type Definitions

Variants